@@ -1,30 +1,42 @@
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/foundation.dart';
+import 'package:stripe_payment/src/stripe_options.dart';
 
 import 'android_pay_payment_request.dart';
 import 'apple_pay_payment_request.dart';
-import 'card_form_payment_request.dart';
 import 'error_codes.dart';
-import 'payment_intent.dart';
-import 'payment_method.dart';
-import 'source.dart';
-import 'source_params.dart';
-import 'token.dart';
+import 'model/bank_account.dart';
+import 'model/card_form_payment_request.dart';
+import 'model/credit_card.dart';
+import 'model/payment_intent.dart';
+import 'model/payment_intent_result.dart';
+import 'model/payment_method.dart';
+import 'model/payment_method_request.dart';
+import 'model/setup_intent_result.dart';
+import 'model/source.dart';
+import 'model/source_params.dart';
+import 'model/token.dart';
 
 class StripePayment {
   static const MethodChannel _channel = const MethodChannel('stripe_payment');
 
   /// https://tipsi.github.io/tipsi-stripe/docs/usage.html
-  static void setOptions(StripeOptions settings) {
-    _channel.invokeMethod('setOptions', {"options": settings.toJson(), "errorCodes": Errors.mapping});
+  static void setOptions(StripeOptions settings) async {
+    await _channel.invokeMethod('setOptions', {
+      "options": settings.toJson(),
+      "errorCodes": Errors.mapping,
+    });
   }
 
   /// https://tipsi.github.io/tipsi-stripe/docs/usage.html
-  static void setStripeAccount(String stripeAccount) {
-    _channel.invokeMethod('setStripeAccount', {"stripeAccount": stripeAccount});
+  static void setStripeAccount(String stripeAccount) async {
+    await _channel.invokeMethod('setStripeAccount', {
+      "stripeAccount": stripeAccount,
+    });
   }
 
   /// https://tipsi.github.io/tipsi-stripe/docs/usage.html
@@ -49,26 +61,69 @@ class StripePayment {
       throw UnimplementedError();
     } else {
       if (Platform.isAndroid) {
-        return _channel.invokeMethod('canMakeAndroidPayPayments');
+        return _channel.invokeMethod<bool>('canMakeAndroidPayPayments');
       } else if (Platform.isIOS) {
         Map<String, dynamic> options = {"networks": networks};
-        return _channel.invokeMethod('canMakeApplePayPayments', options);
+        return _channel.invokeMethod<bool>('canMakeApplePayPayments', options);
       } else
         throw UnimplementedError();
     }
   }
 
   static Future<List<String>> potentiallyAvailableNativePayNetworks() async {
-    return await _channel.invokeListMethod("potentiallyAvailableNativePayNetworks");
+    return await _channel.invokeListMethod<String>(
+      "potentiallyAvailableNativePayNetworks",
+    );
   }
 
-  static Future<bool> _deviceSupportsAndroidPay() => _channel.invokeMethod("deviceSupportsAndroidPay");
+  static Future<bool> _deviceSupportsAndroidPay() {
+    return _channel.invokeMethod<bool>("deviceSupportsAndroidPay");
+  }
 
-  static Future<bool> _deviceSupportsApplePay() => _channel.invokeMethod("deviceSupportsApplePay");
+  static Future<bool> _deviceSupportsApplePay() {
+    return _channel.invokeMethod<bool>("deviceSupportsApplePay");
+  }
+
+  static Future<PaymentMethod> paymentMethodFromNativePay({
+    @required AndroidPayPaymentRequest androidPayOptions,
+    @required ApplePayPaymentOptions applePayOptions,
+  }) {
+    if (kIsWeb) {
+      throw UnimplementedError();
+    } else {
+      if (Platform.isAndroid) {
+        return _paymentMethodFromAndroidPay(androidPayOptions);
+      } else if (Platform.isIOS) {
+        return _paymentMethodFromApplePay(applePayOptions);
+      } else
+        throw UnimplementedError();
+    }
+  }
+
+  static Future<PaymentMethod> _paymentMethodFromAndroidPay(
+      AndroidPayPaymentRequest options) async {
+    final pm = await _channel.invokeMapMethod<dynamic, dynamic>(
+      "paymentMethodFromAndroidPay",
+      options.toJson(),
+    );
+    return PaymentMethod.fromJson(pm);
+  }
+
+  static Future<PaymentMethod> _paymentMethodFromApplePay(
+      ApplePayPaymentOptions options) async {
+    final pm = await _channel
+        .invokeMapMethod<dynamic, dynamic>("paymentMethodFromApplePay", {
+      "options": options.json,
+      "items": options.items.map((item) => item.json).toList()
+    });
+    return PaymentMethod.fromJson(pm);
+  }
 
   /// https://tipsi.github.io/tipsi-stripe/docs/paymentRequestWithNativePay.html
-  static Future<Token> paymentRequestWithNativePay(
-      {@required AndroidPayPaymentRequest androidPayOptions, @required ApplePayPaymentOptions applePayOptions}) {
+  static Future<Token> paymentRequestWithNativePay({
+    @required AndroidPayPaymentRequest androidPayOptions,
+    @required ApplePayPaymentOptions applePayOptions,
+  }) {
     if (kIsWeb) {
       throw UnimplementedError();
     } else {
@@ -81,14 +136,22 @@ class StripePayment {
     }
   }
 
-  static Future<Token> _paymentRequestWithAndroidPay(AndroidPayPaymentRequest options) async {
-    final token = await _channel.invokeMethod("paymentRequestWithAndroidPay", options.toJson());
+  static Future<Token> _paymentRequestWithAndroidPay(
+      AndroidPayPaymentRequest options) async {
+    final token = await _channel.invokeMapMethod<dynamic, dynamic>(
+      "paymentRequestWithAndroidPay",
+      options.toJson(),
+    );
     return Token.fromJson(token);
   }
 
-  static Future<Token> _paymentRequestWithApplePay(ApplePayPaymentOptions options) async {
-    final token = await _channel.invokeMethod("paymentRequestWithApplePay",
-        {"options": options.json, "items": options.items.map((item) => item.json).toList()});
+  static Future<Token> _paymentRequestWithApplePay(
+      ApplePayPaymentOptions options) async {
+    final token = await _channel
+        .invokeMapMethod<dynamic, dynamic>("paymentRequestWithApplePay", {
+      "options": options.json,
+      "items": options.items.map((item) => item.json).toList()
+    });
     return Token.fromJson(token);
   }
 
@@ -98,7 +161,7 @@ class StripePayment {
       throw UnimplementedError();
     } else {
       if (Platform.isIOS) {
-        return _channel.invokeMethod("completeApplePayRequest");
+        return await _channel.invokeMethod("completeApplePayRequest");
       } else if (Platform.isAndroid) {
         return null;
       } else
@@ -112,7 +175,7 @@ class StripePayment {
       throw UnimplementedError();
     } else {
       if (Platform.isIOS) {
-        return _channel.invokeMethod("cancelApplePayRequest");
+        return await _channel.invokeMethod("cancelApplePayRequest");
       } else if (Platform.isAndroid) {
         return null;
       } else
@@ -121,83 +184,94 @@ class StripePayment {
   }
 
   /// https://tipsi.github.io/tipsi-stripe/docs/paymentRequestWithCardForm.html
-  static Future<PaymentMethod> paymentRequestWithCardForm(CardFormPaymentRequest options) async {
-    final token = await _channel.invokeMethod("paymentRequestWithCardForm", options.toJson());
+  static Future<PaymentMethod> paymentRequestWithCardForm(
+      CardFormPaymentRequest options) async {
+    final token = await _channel.invokeMapMethod<dynamic, dynamic>(
+      "paymentRequestWithCardForm",
+      options.toJson(),
+    );
     return PaymentMethod.fromJson(token);
   }
 
   /// https://tipsi.github.io/tipsi-stripe/docs/createTokenWithCard.html
   static Future<Token> createTokenWithCard(CreditCard card) async {
-    final token = await _channel.invokeMethod("createTokenWithCard", card.toJson());
+    final token = await _channel.invokeMapMethod<dynamic, dynamic>(
+      "createTokenWithCard",
+      card.toJson(),
+    );
     return Token.fromJson(token);
   }
 
   /// https://tipsi.github.io/tipsi-stripe/docs/createTokenWithBankAccount.html
   static Future<Token> createTokenWithBankAccount(BankAccount options) async {
-    final token = await _channel.invokeMethod("createTokenWithBankAccount", options.toJson());
+    final token = await _channel.invokeMapMethod<dynamic, dynamic>(
+      "createTokenWithBankAccount",
+      options.toJson(),
+    );
     return Token.fromJson(token);
   }
 
   /// https://tipsi.github.io/tipsi-stripe/docs/createsourcewithparamsparams.html
   static Future<Source> createSourceWithParams(SourceParams options) async {
-    final source = await _channel.invokeMethod("createSourceWithParams", options.toJson());
+    final source = await _channel.invokeMapMethod<dynamic, dynamic>(
+      "createSourceWithParams",
+      options.toJson(),
+    );
     return Source.fromJson(source);
   }
 
   /// https://tipsi.github.io/tipsi-stripe/docs/createPaymentMethod.html
-  static Future<PaymentMethod> createPaymentMethod(PaymentMethodRequest request) async {
-    final paymentMethod = await _channel.invokeMethod("createPaymentMethod", request.toJson());
+  static Future<PaymentMethod> createPaymentMethod(
+      PaymentMethodRequest request) async {
+    final paymentMethod = await _channel.invokeMapMethod<String, dynamic>(
+      "createPaymentMethod",
+      request.toJson(),
+    );
     return PaymentMethod.fromJson(paymentMethod);
   }
 
   /// https://tipsi.github.io/tipsi-stripe/docs/authenticatePaymentIntent.html
-  static Future<PaymentIntentResult> authenticatePaymentIntent({@required String clientSecret}) async {
+  static Future<PaymentIntentResult> authenticatePaymentIntent(
+      {@required String clientSecret}) async {
     assert(clientSecret != null);
-    final result = await _channel.invokeMethod('authenticatePaymentIntent', {"clientSecret": clientSecret});
+    final result = await _channel.invokeMapMethod<dynamic, dynamic>(
+      'authenticatePaymentIntent',
+      {"clientSecret": clientSecret},
+    );
     return PaymentIntentResult.fromJson(result);
   }
 
   /// https://tipsi.github.io/tipsi-stripe/docs/confirmPaymentIntent.html
-  static Future<PaymentIntentResult> confirmPaymentIntent(PaymentIntent intent) async {
+  static Future<PaymentIntentResult> confirmPaymentIntent(
+      PaymentIntent intent) async {
     assert(intent.clientSecret != null);
     assert(intent.paymentMethodId != null);
-    final result = await _channel.invokeMethod('confirmPaymentIntent', intent.toJson());
+    final result = await _channel.invokeMapMethod<dynamic, dynamic>(
+      'confirmPaymentIntent',
+      intent.toJson(),
+    );
     return PaymentIntentResult.fromJson(result);
   }
 
   /// https://tipsi.github.io/tipsi-stripe/docs/authenticateSetupIntent.html
-  static Future<SetupIntentResult> authenticateSetupIntent({@required String clientSecret}) async {
+  static Future<SetupIntentResult> authenticateSetupIntent(
+      {@required String clientSecret}) async {
     assert(clientSecret != null);
-    final result = await _channel.invokeMethod('authenticateSetupIntent', {"clientSecret": clientSecret});
+    final result = await _channel.invokeMapMethod<dynamic, dynamic>(
+      'authenticateSetupIntent',
+      {"clientSecret": clientSecret},
+    );
     return SetupIntentResult.fromJson(result);
   }
 
   /// https://tipsi.github.io/tipsi-stripe/docs/confirmSetupIntent.html
-  static Future<SetupIntentResult> confirmSetupIntent(PaymentIntent intent) async {
+  static Future<SetupIntentResult> confirmSetupIntent(
+      PaymentIntent intent) async {
     assert(intent.clientSecret != null);
-    assert(intent.paymentMethodId != null);
-    final result = await _channel.invokeMethod('confirmSetupIntent', intent.toJson());
+    final result = await _channel.invokeMapMethod<dynamic, dynamic>(
+      'confirmSetupIntent',
+      intent.toJson(),
+    );
     return SetupIntentResult.fromJson(result);
-  }
-}
-
-class StripeOptions {
-  final String publishableKey;
-  final String merchantId;
-  final String androidPayMode;
-
-  StripeOptions({@required this.publishableKey, this.merchantId, this.androidPayMode});
-
-  factory StripeOptions.fromJson(Map<String, dynamic> json) {
-    return StripeOptions(
-        merchantId: json['merchantId'], publishableKey: json['publishableKey'], androidPayMode: json['androidPayMode']);
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    if (this.merchantId != null) data['merchantId'] = this.merchantId;
-    if (this.publishableKey != null) data['publishableKey'] = this.publishableKey;
-    if (this.androidPayMode != null) data['androidPayMode'] = this.androidPayMode;
-    return data;
   }
 }
