@@ -1,9 +1,20 @@
 package com.gettipsi.stripe;
 
+import static com.gettipsi.stripe.Errors.toErrorCode;
+import static com.gettipsi.stripe.util.Converters.convertTokenToWritableMap;
+import static com.gettipsi.stripe.util.Converters.getAllowedShippingCountryCodes;
+import static com.gettipsi.stripe.util.Converters.getBillingAddress;
+import static com.gettipsi.stripe.util.Converters.putExtraToTokenMap;
+import static com.gettipsi.stripe.util.PayParams.BILLING_ADDRESS_REQUIRED;
+import static com.gettipsi.stripe.util.PayParams.CURRENCY_CODE;
+import static com.gettipsi.stripe.util.PayParams.EMAIL_REQUIRED;
+import static com.gettipsi.stripe.util.PayParams.PHONE_NUMBER_REQUIRED;
+import static com.gettipsi.stripe.util.PayParams.SHIPPING_ADDRESS_REQUIRED;
+import static com.gettipsi.stripe.util.PayParams.TOTAL_PRICE;
+
 import android.app.Activity;
 import android.content.Intent;
 import androidx.annotation.NonNull;
-
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReadableMap;
 import com.gettipsi.stripe.util.ArgCheck;
@@ -11,7 +22,6 @@ import com.gettipsi.stripe.util.Converters;
 import com.gettipsi.stripe.util.Fun0;
 import com.google.android.gms.common.api.ApiException;
 import com.google.android.gms.common.api.Status;
-import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.android.gms.wallet.AutoResolveHelper;
 import com.google.android.gms.wallet.CardRequirements;
@@ -26,26 +36,12 @@ import com.google.android.gms.wallet.Wallet;
 import com.google.android.gms.wallet.WalletConstants;
 import com.stripe.android.BuildConfig;
 import com.stripe.android.model.Token;
-
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.List;
 
-import static com.gettipsi.stripe.Errors.toErrorCode;
-import static com.gettipsi.stripe.util.Converters.convertTokenToWritableMap;
-import static com.gettipsi.stripe.util.Converters.getAllowedShippingCountryCodes;
-import static com.gettipsi.stripe.util.Converters.getBillingAddress;
-import static com.gettipsi.stripe.util.Converters.putExtraToTokenMap;
-import static com.gettipsi.stripe.util.PayParams.CURRENCY_CODE;
-import static com.gettipsi.stripe.util.PayParams.BILLING_ADDRESS_REQUIRED;
-import static com.gettipsi.stripe.util.PayParams.SHIPPING_ADDRESS_REQUIRED;
-import static com.gettipsi.stripe.util.PayParams.PHONE_NUMBER_REQUIRED;
-import static com.gettipsi.stripe.util.PayParams.EMAIL_REQUIRED;
-import static com.gettipsi.stripe.util.PayParams.TOTAL_PRICE;
-
-/**
- * Created by ngoriachev on 13/03/2018.
- * see https://developers.google.com/pay/api/tutorial
- */
+/** Created by ngoriachev on 13/03/2018. see https://developers.google.com/pay/api/tutorial */
 public final class GoogleApiPayFlowImpl extends PayFlow {
 
   private static final String TAG = GoogleApiPayFlowImpl.class.getSimpleName();
@@ -60,135 +56,137 @@ public final class GoogleApiPayFlowImpl extends PayFlow {
 
   private PaymentsClient createPaymentsClient(@NonNull Activity activity) {
     return Wallet.getPaymentsClient(
-      activity,
-      new Wallet.WalletOptions.Builder().setEnvironment(getEnvironment()).build());
+        activity, new Wallet.WalletOptions.Builder().setEnvironment(getEnvironment()).build());
   }
 
-  private void isReadyToPay(@NonNull Activity activity, boolean isExistingPaymentMethodRequired, @NonNull final Promise promise) {
+  private void isReadyToPay(
+      @NonNull Activity activity,
+      boolean isExistingPaymentMethodRequired,
+      @NonNull final Promise promise) {
     ArgCheck.nonNull(activity);
     ArgCheck.nonNull(promise);
 
     IsReadyToPayRequest request =
-      IsReadyToPayRequest.newBuilder()
-        .addAllowedPaymentMethod(WalletConstants.PAYMENT_METHOD_CARD)
-        .addAllowedPaymentMethod(WalletConstants.PAYMENT_METHOD_TOKENIZED_CARD)
-        .setExistingPaymentMethodRequired(isExistingPaymentMethodRequired)
-        .build();
+        IsReadyToPayRequest.newBuilder()
+            .addAllowedPaymentMethod(WalletConstants.PAYMENT_METHOD_CARD)
+            .addAllowedPaymentMethod(WalletConstants.PAYMENT_METHOD_TOKENIZED_CARD)
+            .setExistingPaymentMethodRequired(isExistingPaymentMethodRequired)
+            .build();
     mPaymentsClient = createPaymentsClient(activity);
     Task<Boolean> task = mPaymentsClient.isReadyToPay(request);
     task.addOnCompleteListener(
-      new OnCompleteListener<Boolean>() {
-        public void onComplete(Task<Boolean> task) {
+        task1 -> {
           try {
-            boolean result = task.getResult(ApiException.class);
+            Boolean result = task1.getResult(ApiException.class);
             promise.resolve(result);
           } catch (ApiException exception) {
             promise.reject(toErrorCode(exception), exception.getMessage());
           }
-        }
-      });
+        });
   }
 
   private PaymentMethodTokenizationParameters createPaymentMethodTokenizationParameters() {
     return PaymentMethodTokenizationParameters.newBuilder()
-      .setPaymentMethodTokenizationType(WalletConstants.PAYMENT_METHOD_TOKENIZATION_TYPE_PAYMENT_GATEWAY)
-      .addParameter("gateway", "stripe")
-      .addParameter("stripe:publishableKey", getPublishableKey())
-      .addParameter("stripe:version", BuildConfig.VERSION_NAME)
-      .build();
+        .setPaymentMethodTokenizationType(
+            WalletConstants.PAYMENT_METHOD_TOKENIZATION_TYPE_PAYMENT_GATEWAY)
+        .addParameter("gateway", "stripe")
+        .addParameter("stripe:publishableKey", getPublishableKey())
+        .addParameter("stripe:version", BuildConfig.VERSION_NAME)
+        .build();
   }
 
   private PaymentDataRequest createPaymentDataRequest(ReadableMap payParams) {
     final String estimatedTotalPrice = payParams.getString(TOTAL_PRICE);
     final String currencyCode = payParams.getString(CURRENCY_CODE);
-    final boolean billingAddressRequired = Converters.getValue(payParams, BILLING_ADDRESS_REQUIRED, false);
-    final boolean shippingAddressRequired = Converters.getValue(payParams, SHIPPING_ADDRESS_REQUIRED, false);
-    final boolean phoneNumberRequired = Converters.getValue(payParams, PHONE_NUMBER_REQUIRED, false);
+    final boolean billingAddressRequired =
+        Converters.getValue(payParams, BILLING_ADDRESS_REQUIRED, false);
+    final boolean shippingAddressRequired =
+        Converters.getValue(payParams, SHIPPING_ADDRESS_REQUIRED, false);
+    final boolean phoneNumberRequired =
+        Converters.getValue(payParams, PHONE_NUMBER_REQUIRED, false);
     final boolean emailRequired = Converters.getValue(payParams, EMAIL_REQUIRED, false);
     final Collection<String> allowedCountryCodes = getAllowedShippingCountryCodes(payParams);
 
     return createPaymentDataRequest(
-      estimatedTotalPrice,
-      currencyCode,
-      billingAddressRequired,
-      shippingAddressRequired,
-      phoneNumberRequired,
-      emailRequired,
-      allowedCountryCodes
-    );
+        estimatedTotalPrice,
+        currencyCode,
+        billingAddressRequired,
+        shippingAddressRequired,
+        phoneNumberRequired,
+        emailRequired,
+        allowedCountryCodes);
   }
 
-  private PaymentDataRequest createPaymentDataRequest(@NonNull final String totalPrice,
-                                                      @NonNull final String currencyCode,
-                                                      final boolean billingAddressRequired,
-                                                      final boolean shippingAddressRequired,
-                                                      final boolean phoneNumberRequired,
-                                                      final boolean emailRequired,
-                                                      @NonNull final Collection<String> countryCodes
-  ) {
+  private PaymentDataRequest createPaymentDataRequest(
+      @NonNull final String totalPrice,
+      @NonNull final String currencyCode,
+      final boolean billingAddressRequired,
+      final boolean shippingAddressRequired,
+      final boolean phoneNumberRequired,
+      final boolean emailRequired,
+      @NonNull final Collection<String> countryCodes) {
 
     ArgCheck.isDouble(totalPrice);
     ArgCheck.notEmptyString(currencyCode);
 
     PaymentDataRequest.Builder builder = PaymentDataRequest.newBuilder();
     builder.setTransactionInfo(
-      TransactionInfo.newBuilder()
-        .setTotalPriceStatus(WalletConstants.TOTAL_PRICE_STATUS_ESTIMATED)
-        .setTotalPrice(totalPrice)
-        .setCurrencyCode(currencyCode)
-        .build());
+        TransactionInfo.newBuilder()
+            .setTotalPriceStatus(WalletConstants.TOTAL_PRICE_STATUS_ESTIMATED)
+            .setTotalPrice(totalPrice)
+            .setCurrencyCode(currencyCode)
+            .build());
 
     builder
-      .setCardRequirements(
-        CardRequirements.newBuilder()
-          .addAllowedCardNetworks(
-            Arrays.asList(
-              WalletConstants.CARD_NETWORK_AMEX,
-              WalletConstants.CARD_NETWORK_DISCOVER,
-              WalletConstants.CARD_NETWORK_VISA,
-              WalletConstants.CARD_NETWORK_MASTERCARD))
-          .setBillingAddressRequired(billingAddressRequired)
-          .build())
-      .addAllowedPaymentMethod(WalletConstants.PAYMENT_METHOD_CARD)
-      .addAllowedPaymentMethod(WalletConstants.PAYMENT_METHOD_TOKENIZED_CARD)
-      .setEmailRequired(emailRequired)
-      .setShippingAddressRequired(shippingAddressRequired)
-      .setPhoneNumberRequired(phoneNumberRequired);
+        .setCardRequirements(
+            CardRequirements.newBuilder()
+                .addAllowedCardNetworks(
+                    Arrays.asList(
+                        WalletConstants.CARD_NETWORK_AMEX,
+                        WalletConstants.CARD_NETWORK_DISCOVER,
+                        WalletConstants.CARD_NETWORK_JCB,
+                        WalletConstants.CARD_NETWORK_MASTERCARD,
+                        WalletConstants.CARD_NETWORK_VISA,
+                        WalletConstants.CARD_NETWORK_INTERAC,
+                        WalletConstants.CARD_NETWORK_OTHER))
+                .setBillingAddressRequired(billingAddressRequired)
+                .build())
+        .addAllowedPaymentMethod(WalletConstants.PAYMENT_METHOD_CARD)
+        .addAllowedPaymentMethod(WalletConstants.PAYMENT_METHOD_TOKENIZED_CARD)
+        .setEmailRequired(emailRequired)
+        .setShippingAddressRequired(shippingAddressRequired)
+        .setPhoneNumberRequired(phoneNumberRequired);
 
     if (countryCodes.size() > 0) {
       builder.setShippingAddressRequirements(
-        ShippingAddressRequirements.newBuilder()
-          .addAllowedCountryCodes(countryCodes)
-          .build());
+          ShippingAddressRequirements.newBuilder().addAllowedCountryCodes(countryCodes).build());
     }
 
     builder.setPaymentMethodTokenizationParameters(createPaymentMethodTokenizationParameters());
     return builder.build();
   }
 
-  private void startPaymentRequest(@NonNull Activity activity, @NonNull PaymentDataRequest request) {
+  private void startPaymentRequest(
+      @NonNull Activity activity, @NonNull PaymentDataRequest request) {
     ArgCheck.nonNull(activity);
     ArgCheck.nonNull(request);
 
     mPaymentsClient = createPaymentsClient(activity);
 
     AutoResolveHelper.resolveTask(
-      mPaymentsClient.loadPaymentData(request),
-      activity,
-      LOAD_PAYMENT_DATA_REQUEST_CODE);
+        mPaymentsClient.loadPaymentData(request), activity, LOAD_PAYMENT_DATA_REQUEST_CODE);
   }
 
   @Override
-  public void paymentRequestWithAndroidPay(@NonNull ReadableMap payParams, @NonNull Promise promise) {
+  public void paymentRequestWithAndroidPay(
+      @NonNull ReadableMap payParams, @NonNull Promise promise) {
     ArgCheck.nonNull(payParams);
     ArgCheck.nonNull(promise);
 
     Activity activity = activityProvider.call();
     if (activity == null) {
       promise.reject(
-        getErrorCode("activityUnavailable"),
-        getErrorDescription("activityUnavailable")
-      );
+          getErrorCode("activityUnavailable"), getErrorDescription("activityUnavailable"));
       return;
     }
 
@@ -197,25 +195,30 @@ public final class GoogleApiPayFlowImpl extends PayFlow {
   }
 
   @Override
-  public void deviceSupportsAndroidPay(boolean isExistingPaymentMethodRequired, @NonNull Promise promise) {
+  public void deviceSupportsAndroidPay(
+      boolean isExistingPaymentMethodRequired, @NonNull Promise promise) {
     Activity activity = activityProvider.call();
     if (activity == null) {
       promise.reject(
-        getErrorCode("activityUnavailable"),
-        getErrorDescription("activityUnavailable")
-      );
+          getErrorCode("activityUnavailable"), getErrorDescription("activityUnavailable"));
       return;
     }
 
     if (!isPlayServicesAvailable(activity)) {
       promise.reject(
-        getErrorCode("playServicesUnavailable"),
-        getErrorDescription("playServicesUnavailable")
-      );
+          getErrorCode("playServicesUnavailable"), getErrorDescription("playServicesUnavailable"));
       return;
     }
 
     isReadyToPay(activity, isExistingPaymentMethodRequired, promise);
+  }
+
+  public void potentiallyAvailableNativePayNetworks(@NonNull Promise promise) {
+    List<String> networks = new ArrayList<>();
+    networks.add("visa");
+    networks.add("jcb");
+    networks.add("master");
+    promise.resolve(networks);
   }
 
   public boolean onActivityResult(Activity activity, int requestCode, int resultCode, Intent data) {
@@ -233,32 +236,26 @@ public final class GoogleApiPayFlowImpl extends PayFlow {
             Token token = Token.fromString(tokenJson);
             if (token == null) {
               payPromise.reject(
-                getErrorCode("parseResponse"),
-                getErrorDescription("parseResponse")
-              );
+                  getErrorCode("parseResponse"), getErrorDescription("parseResponse"));
             } else {
-              payPromise.resolve(putExtraToTokenMap(
-                convertTokenToWritableMap(token),
-                getBillingAddress(paymentData),
-                paymentData.getShippingAddress(),
-                paymentData.getEmail()));
+              payPromise.resolve(
+                  putExtraToTokenMap(
+                      convertTokenToWritableMap(token),
+                      getBillingAddress(paymentData),
+                      paymentData.getShippingAddress(),
+                      paymentData.getEmail()));
             }
             break;
           case Activity.RESULT_CANCELED:
             payPromise.reject(
-              getErrorCode("purchaseCancelled"),
-              getErrorDescription("purchaseCancelled")
-            );
+                getErrorCode("purchaseCancelled"), getErrorDescription("purchaseCancelled"));
             break;
           case AutoResolveHelper.RESULT_ERROR:
             Status status = AutoResolveHelper.getStatusFromIntent(data);
             // Log the status for debugging.
             // Generally, there is no need to show an error to
             // the user as the Google Pay API will do that.
-            payPromise.reject(
-              getErrorCode("stripe"),
-              status.getStatusMessage()
-            );
+            payPromise.reject(getErrorCode("stripe"), status.getStatusMessage());
             break;
 
           default:
@@ -270,5 +267,4 @@ public final class GoogleApiPayFlowImpl extends PayFlow {
 
     return false;
   }
-
 }
