@@ -1,14 +1,10 @@
 package com.gettipsi.stripe.util;
 
-/**
- * Created by ngoriachev on 15/03/2018.
- */
-
+/** Created by ngoriachev on 15/03/2018. */
 public abstract class InitializationOptions {
 
   public static final String PUBLISHABLE_KEY = "publishableKey";
   public static final String ANDROID_PAY_MODE_KEY = "androidPayMode";
   public static final String ANDROID_PAY_MODE_PRODUCTION = "production";
   public static final String ANDROID_PAY_MODE_TEST = "test";
-
 }
